refactor: HandlerAdapters 일급 객체화
